backend/requirements.txt
accelerate==1.7.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
datasets==3.6.0
dill==0.3.8
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
h11==0.16.0
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kagglehub==0.3.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
networkx==3.5
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
pyarrow==20.0.0
pycocotools==2.0.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
